modified banner

+++ src/Pages/MyCampaign.jsx
@@ -17,22 +17,38 @@ const MyCampaign = (props) => {
     }, [])
 
     const handleDelete = (id) => {
-        console.log(id)
-        fetch(`http://localhost:5000/delete/${id}`,
-            { method: "DELETE" })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                if (data.deletedCount > 0) {
-                    Swal.fire({
-                        title: "Campaign Updated!",
-                        icon: "success"
-                    });}
+        
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
 
-                const newCampaign = campaign.filter(c=> c._id !==  id);
-                setCampaign(newCampaign)
+                fetch(`http://localhost:5000/delete/${id}`,
+                    { method: "DELETE" })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                title: "Deleted",
+                                icon: "success"
+                            });}
+        
+                        const newCampaign = campaign.filter(c=> c._id !==  id);
+                        setCampaign(newCampaign)
+        
+                    })}
+          });
 
-            })
+
+
+        
     }
 
     return (
